users can now update their informations and view their orders

+++ src/app/pages/checkout/checkout.page.ts
@@ -5,10 +5,13 @@ import {ToastService} from '../../services/toast.service';
 import {AlertController} from '@ionic/angular';
 import {Router} from '@angular/router';
 import {LoadingService} from '../../services/loading.service';
-import {AuthenticationService} from '../../services/authentication.service';
+import {UserManagementService} from '../../services/user-management.service';
 import {UserData} from '../../objects/user-data';
 import {DarkModeService} from '../../services/dark-mode.service';
 import {UnsubscribeOnDestroyAdapter} from '../../utilities/unsubscribe-on-destroy-adapter';
+import {HttpClient} from '@angular/common/http';
+import {UserOrders} from '../../objects/user-orders';
+import {Order} from '../../objects/order';
 
 @Component({
     selector: 'app-checkout',
@@ -26,8 +29,8 @@ export class CheckoutPage extends UnsubscribeOnDestroyAdapter implements OnInit 
     public headerColor: string;
 
     constructor(private cartService: CartService, private alertController: AlertController, private toastService: ToastService,
-                private router: Router, private loadingService: LoadingService, private authService: AuthenticationService,
-                private darkModeService: DarkModeService) {
+                private router: Router, private loadingService: LoadingService, private userManagementService: UserManagementService,
+                private darkModeService: DarkModeService, private httpClient: HttpClient) {
         super();
 
         this.products = [];
@@ -45,12 +48,13 @@ export class CheckoutPage extends UnsubscribeOnDestroyAdapter implements OnInit 
             });
         }));
 
-        this.subscriptions.add(this.authService.getUser().subscribe(user => {
+        this.subscriptions.add(this.userManagementService.getUser().subscribe(user => {
             this.user = user;
         }));
     }
 
-    ngOnInit() {}
+    ngOnInit() {
+    }
 
     async sendInquiry() {
         const alert = await this.alertController.create({
@@ -65,15 +69,28 @@ export class CheckoutPage extends UnsubscribeOnDestroyAdapter implements OnInit 
                     handler: async () => {
                         await this.loadingService.showLoading();
 
-                        // TODO send inquiry
+                        const order = {} as UserOrders;
+                        order.userId = Number(this.user.id);
+                        order.orders = this.products.map(value => {
+                            return {
+                                productId: Number(value.getProduct().getId()),
+                                amount: value.getQuantity()
+                            } as Order;
+                        });
 
-                        await this.cartService.removeAllProductsFromCart();
-                        await this.router.navigate(['/products']);
+                        const response = await this.httpClient.post<OrderResponse>('https://inn3-webshop.lukas-hinterleitner.at/api/orders/',
+                            order, {}).toPromise();
+
+                        const success = response.success;
+
+                        if (response.success === 1) {
+                            await this.cartService.removeAllProductsFromCart();
+                            await this.router.navigate(['user/orders', response.orderNumber]);
+                        }
 
                         await this.loadingService.closeLoading();
-
-                        await this.toastService.showToast(2500, '', 'Your inquiry was sent to us!' +
-                            ' Thank you for shopping with us.', 'success');
+                        await this.toastService.showToast(2500, '', success ? 'Your inquiry was sent to us!' +
+                            ' Thank you for shopping with us.' : response.message, success ? 'success' : 'danger');
                     }
                 }
             ]
@@ -81,4 +98,12 @@ export class CheckoutPage extends UnsubscribeOnDestroyAdapter implements OnInit 
 
         await alert.present();
     }
+}
+
+
+export interface OrderResponse {
+    success: number;
+    status: number;
+    message: string;
+    orderNumber: string;
 }
